remove duplicated code by introducing upper level DebugOptimizer class
